requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.0
DateTime==5.5
distro==1.9.0
gobject==0.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.61.0
packaging==24.2
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
pytz==2025.1
pywebview==5.3.2
QtPy==2.4.2
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
zope.interface==7.2
